horus/requirements.txt
psycopg2-binary
sqlalchemy
pandas
numpy
matplotlib
pydantic
websockets
python-dotenv
